+++ src/patterns/notTriplePatterns/UnionPattern.spec.ts
@@ -57,7 +57,7 @@ describe( "UnionPattern", () => {
 				and: jasmine.any( Function ),
 
 				// Inherit
-				getPattern: jasmine.any( Function ),
+				_getPattern: jasmine.any( Function ),
 			} );
 		} );
 
@@ -82,7 +82,7 @@ describe( "UnionPattern", () => {
 			expect( returned ).toEqual( {
 				and: jasmine.any( Function ),
 
-				getPattern: jasmine.any( Function ),
+				_getPattern: jasmine.any( Function ),
 			} )
 		} );